backend/pyproject.toml
[tool.poetry]
name = "kubota-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kubota_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
drf-yasg = "^1.21.4"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^22.1"
gunicorn = "^20.1.0"
django-cors-headers = "^3.13.0"
whitenoise = "^6.2.0"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
